# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "NcrRugby API",
    "description": "NcrRugby spec for Swagger",
    "termsOfService": "http://swagger.io/terms/"
  },
  "basePath": "/api/v1",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "multipart/form-data"
  ],
  "produces": [
    "application/json"
  ],
  "paths":{
...
    "/getInCompleteBracketUsers": {
      "get": {
        "tags": [
          "User Bracket"
        ],
        "description": "getInCompleteBracketUsers",
        "operationId": "getInCompleteBracketUsers",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name":"access_token",
            "in":"header",
            "description":"access_token",
            "required":false,
            "type":"string",
            "default":""
         }
        ],
        "responses": {
          "200": {
            "description": "Incomplete Bracket Details Found successfully"
          }
        }
      }
    },
...
    },
...
}